// ui/src/lib/types/api.types.ts
export interface components {
    schemas: {
// ...
        CalculationInputDto: {
            amount: number;
            rate?: number;
            term?: number;
            /** Format: date */
            startDate?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}